Cloud Functions/MLP-Test/requirements.txt
numpy
google-cloud-storage
sklearn